update spec with bigger values. better error handling for edges. update rust lib accordingly.

+++ v0/alcov.h
@@ -40,8 +40,8 @@ struct __attribute__((packed)) alcov_hdr {
 	uint64_t magic; 					// equals ALCOV_MAGIC, always the same across every version.
 	uint64_t version_major; 			// equals ALCOV_VERSION_MAJOR, increases when the specification changes in a significant way.
 	uint64_t version_minor; 			// equals ALCOV_VERSION_MINOR, increases when the specification changes include minor breaking changes.
-	uint32_t nb_modules;				// number of blocks used during coverage.
-	uint32_t nb_blocks; 				// number of modules.
+	uint16_t nb_modules;				// number of blocks used during coverage.
+	uint64_t nb_blocks; 				// number of modules.
 	uint64_t nb_edges;					// number of edges.
 	uint64_t modules_start;				// offset of modules chunk in file.
 	uint64_t paths_start;				// offset of paths chunk in file.
@@ -58,15 +58,15 @@ struct __attribute__((packed)) alcov_segment {
 struct __attribute__((packed)) alcov_module {
 	uint64_t base_address;				// base address of the module.
 	int64_t path_offset;				// offset (in bytes) of the path from paths_start. < 0 if no path is provided.
-	uint8_t nb_segments;				// number of segments in next array. must be at least 1.
+	uint16_t nb_segments;				// number of segments in next array. must be at least 1.
 	struct alcov_segment segments[];	// Segments in the modules.
 };
 
 struct __attribute__((packed)) alcov_block {
-	uint32_t segment_offset;			// the block offset in its segment.
-	uint16_t size;						// the size of the block.
-	uint8_t module_id;					// the module ID in which the block lives.
-	uint8_t segment_id;					// the segment ID in which the block lives.
+	uint64_t segment_offset;			// the block offset in its segment.
+	uint32_t size;						// the size of the block.
+	uint16_t module_id;					// the module ID in which the block lives.
+	uint16_t segment_id;				// the segment ID in which the block lives.
 	uint64_t nb_out_edges;				// number of outgoing edges. in the block's edge table. 0 if no outgoing edges.
 	uint64_t out_edges_offset; 			// the offset (in bytes) in the outgoing edge table. only defined when the EDGES flag is set and nb_out_edges > 0.
 	uint64_t nb_taken;					// the number of times the block has been traversed. 0 means it was not measured and this number is unknown.
